+++ models/liu2019/experiments/reproduce-results/run.sh
@@ -4,42 +4,45 @@ DEVICE=1
 
 repro predict \
   --model-name liu2019-transformerabs \
-  --model-args '{"device": '${DEVICE}'}' \
+  --model-kwargs '{"device": '${DEVICE}'}' \
   --dataset-name cnn_dailymail/3.0.0 \
   --split test \
   --output-writer sacrerouge \
-  --output-file ${DIR}/output/cnn_dailymail/transformerabs/predictions.jsonl
+  --output ${DIR}/output/cnn_dailymail/transformerabs/predictions.jsonl
 
 repro predict \
   --model-name liu2019-bertsumext \
-  --model-args '{"device": '${DEVICE}'}' \
+  --model-kwargs '{"device": '${DEVICE}'}' \
   --dataset-name cnn_dailymail/3.0.0 \
   --split test \
   --output-writer sacrerouge \
-  --output-file ${DIR}/output/cnn_dailymail/bertsumext/predictions.jsonl
+  --output ${DIR}/output/cnn_dailymail/bertsumext/predictions.jsonl
 
 repro predict \
   --model-name liu2019-bertsumextabs \
-  --model-args '{"device": '${DEVICE}'}' \
+  --model-kwargs '{"device": '${DEVICE}'}' \
   --dataset-name cnn_dailymail/3.0.0 \
   --split test \
   --output-writer sacrerouge \
-  --output-file ${DIR}/output/cnn_dailymail/bertsumextabs/predictions.jsonl
+  --output ${DIR}/output/cnn_dailymail/bertsumextabs/predictions.jsonl
 
 repro predict \
   --model-name liu2019-bertsumextabs \
-  --model-args '{"device": '${DEVICE}', "model": "bertsumextabs_xsum.pt"}' \
+  --model-kwargs '{"device": '${DEVICE}', "model": "bertsumextabs_xsum.pt"}' \
   --dataset-name xsum/1.2.0 \
   --split test \
   --output-writer sacrerouge \
-  --output-file ${DIR}/output/xsum/bertsumextabs/predictions.jsonl
+  --output ${DIR}/output/xsum/bertsumextabs/predictions.jsonl
 
 for model in "transformerabs" "bertsumext" "bertsumextabs"; do
   for dataset in "cnn_dailymail" "xsum"; do
     if [ -f ${DIR}/output/${dataset}/${model}/predictions.jsonl ]; then
-      python models/sacrerouge/scripts/calculate_rouge.py \
-        --input-file ${DIR}/output/${dataset}/${model}/predictions.jsonl \
-        --output-file ${DIR}/output/${dataset}/${model}/rouge.json
+      repro predict \
+        --model-name sacrerouge-rouge \
+        --input-files ${DIR}/output/${dataset}/${model}/predictions.jsonl \
+        --dataset-reader sacrerouge \
+        --output-writer metrics \
+        --output ${DIR}/output/${dataset}/${model}/rouge.json
     fi
   done
 done