+++ repro/models/squad_v2.py
@@ -1,0 +1,2 @@
+from ._models.squad_v2.src.model import SQuADv2Evaluation
+from ._models.squad_v2.src.setup import SQuADv2SetupSubcommand